Sunmin's data - ELS 2002 (data via e-mail)
Research question: Build a prediction model to find out key features of drop-out students in K-12. 

1. Description of data: 
The data that I selected for my research question is Education Longitudinal Study of 2002 (ELS: 2000) provided from National Center for Education Statistics (NCES). ELS: 2000 is a nationally representative, longitudinal data of 10th graders in 2002 and 12th graders in 2004 who followed throughout the secondary and post-secondary years. It includes surveys from students, parents, Math and English teachers, and school administrators as well as student assessments in Math and English. High school transcripts are also available but restricted. NCES conducted a study in the beginning of 2002 for high school sophomores, first following up in 2004, second in 2006, and final in 2012 (NCES, n.d.). I chose this dataset since it was providing questions related to reasons for drop-out students. 183 variables were identified to be closely associated wi